refator: setting tab
